Rename files to support Vercel
