+++ vercel.json
@@ -1,0 +1,3 @@
+{
+    "rewrites": [{ "source": "/(.*)", "destination": "/api/index" }]
+}
